part-03-python-agent/requirements.txt
# Part 3 specific dependencies
agent-framework>=0.5.0
azure-identity>=1.15.0
pydantic>=2.5.0
python-dotenv>=1.0.0
